+++ core/interfaces.go
@@ -1,6 +1,8 @@
 package core
 
 import (
+	"core-building-block/core/model"
+
 	log "github.com/rokmetro/logging-library/loglib"
 )
 
@@ -31,6 +33,7 @@ func (s *servicesImpl) SerGetCommonTest(l *log.Log) string {
 type Administration interface {
 	AdmGetTest() string
 	AdmGetTestModel() string
+	AdmCreateGlobalConfig(setting string) (*model.GlobalConfig, error)
 }
 
 type administrationImpl struct {
@@ -43,6 +46,10 @@ func (s *administrationImpl) AdmGetTest() string {
 
 func (s *administrationImpl) AdmGetTestModel() string {
 	return s.app.admGetTestModel()
+}
+
+func (s *administrationImpl) AdmCreateGlobalConfig(setting string) (*model.GlobalConfig, error) {
+	return s.app.admCreateGlobalConfig(setting)
 }
 
 //Encryption exposes APIs for the Encryption building block
@@ -74,6 +81,7 @@ func (s *bbsImpl) BBsGetTest() string {
 //Storage is used by core to storage data - DB storage adapter, file storage adapter etc
 type Storage interface {
 	SetStorageListener(storageListener StorageListener)
+	CreateGlobalConfig(setting string) (*model.GlobalConfig, error)
 }
 
 //StorageListener listenes for change data storage events