+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Set up auth framework. [#16] (https://github.com/rokwire/core-building-block/issues/16)
 - Expose version API. [#13] (https://github.com/rokwire/core-building-block/issues/13)
-- Set up project skeleton. [#1](https://github.com/rokwire/core-building-block/issues/1)
-- Define data model. [#2](https://github.com/rokwire/core-building-block/issues/2)
-- Set up logging [#6](https://github.com/rokwire/core-building-block/issues/6)
+- Set up project skeleton. [#1] (https://github.com/rokwire/core-building-block/issues/1)
+- Define data model. [#2] (https://github.com/rokwire/core-building-block/issues/2)
+- Expose create global config admin API [#34] (https://github.com/rokwire/core-building-block/issues/34)
+- Set up logging [#6] (https://github.com/rokwire/core-building-block/issues/6)